requirements.txt
asgiref
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
distlib==0.3.4
Django==4.1
django-axes==5.39.0
django-ipware==4.0.2
django-paypal==1.1.2
django-rq==2.5.1
filelock==3.6.0
google-api-core==2.4.0
google-api-python-client==2.35.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
install==1.3.5
oauthlib==3.1.1
packaging==21.3
pbr==5.8.1
pep8==1.7.1
pip-autoremove==0.10.0
platformdirs==2.5.1
protobuf==3.19.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.1
requests==2.27.1
requests-oauthlib==1.3.0
rq==1.10.1
rsa==4.8
sentry-sdk==1.5.4
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.14.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wrapt==1.13.3
